Move compare_graphs function to tests.helpers module
Because it is not a fixture, it does not belong in conftest.py
