+++ tests/helpers.py
@@ -3,10 +3,12 @@
 # Standard Library
 import collections.abc
 import contextlib
+import json
 
 # Third-party
 import pydantic_settings
 import rdflib
+import rdflib.compare
 
 # Local
 from abis_mapping import settings
@@ -71,3 +73,61 @@ class TestSettings(settings._Settings):
         # This is so the test suite is deterministic and isolated,
         # and won't be effected by any settings a particular developer has set locally.
         return (init_settings,)
+
+
+def compare_graphs(
+    graph1: rdflib.Graph | str,
+    graph2: rdflib.Graph | str,
+) -> bool:
+    """Isomorphically compares to graphs for equality.
+
+    Args:
+        graph1: Graph or Turtle String to Compare
+        graph2: Graph or Turtle String to Compare
+
+    Returns:
+        Whether the graphs are isomorphically equivalent.
+    """
+    # Serialize Graphs if Applicable
+    # There appears to be a difference between the handling of blank-nodes in
+    # graphs *constructed* programmatically by `rdflib`, and graphs *parsed* by
+    # `rdflib`. As such, an easy work-around for testing is to do a round trip
+    # of serialization and de-serialization before the isomorphic comparison.
+    if isinstance(graph1, rdflib.Graph):
+        graph1 = graph1.serialize(format="text/turtle")
+    if isinstance(graph2, rdflib.Graph):
+        graph2 = graph2.serialize(format="text/turtle")
+
+    # Re-Parse Graphs
+    graph1 = rdflib.Graph().parse(data=graph1, format="text/turtle")
+    graph2 = rdflib.Graph().parse(data=graph2, format="text/turtle")
+
+    # Asserts
+    assert isinstance(graph1, rdflib.Graph)
+    assert isinstance(graph2, rdflib.Graph)
+
+    for graph in (graph1, graph2):
+        for s, p, o in graph.triples((None, None, None)):
+            if isinstance(o, rdflib.Literal):
+                # Replace Timestamps
+                # In many cases, dates and datetimes are generately systematically as a
+                # timestamp for "now". When unit testing, we don't care if "now" has
+                # changed when comparing graphs. As such, we want to replace all literals
+                # with a datatype `xsd:date`, `xsd:dateTime` or `xsd:dateTimeStamp` with a
+                # pre-generate value.
+                if o.datatype in (rdflib.XSD.date, rdflib.XSD.dateTime, rdflib.XSD.dateTimeStamp):
+                    graph.set((s, p, rdflib.Literal("test-value")))
+                # Reformat JSON strings
+                # Since the ordering of json keys could be in different ordering depending on
+                # serializer, need to deserialize json then reserialize to ensure the same format
+                # string literal for each.
+                if o.datatype == rdflib.RDF.JSON:
+                    o_dict = json.loads(str(o))
+                    sorted_string = json.dumps(o_dict, sort_keys=True)
+                    graph.set((s, p, rdflib.Literal(sorted_string, datatype=rdflib.RDF.JSON)))
+
+    # Compare Graphs
+    return rdflib.compare.isomorphic(
+        graph1=graph1,
+        graph2=graph2,
+    )
